POC: rtc running on battery
Uses a semaphore and a mutex to sync access to the SPI bus. This seems to be the easiest solution to make the Rust compiler happy, although it is not really necessary, as sd card, RTC is never used at the same time on the different cores. But the overhead is minimal, as the critical_section is only used during aquiring and freeing the mutex.