Indicate with a red LED while savegame is stored
In order to give a reference to the savegame task on the second core the debug led task must go, as there can be only one task to own the reference. Debug LED is now driven by the bootloader implementation which is dropped when it's finished. This makes also sense for later on, when the LED tester is implemented.