Bug: LED does not always receive correct values
On some hardware it seems that the LED does not always receive the correct value. I'm not totally sure what caused this as as soon I put on the probe to measure the signal it's working. So the probe fixes the signal. The input buffer was enabled on the pin, maybe it's related to E9 bug. But that alone did not help, enabling the pull down was also necessary. This will probably improve the high to low level transition time and fixes the issue.

Also put the the part that moves the values into the SPI FIFO inside a critical section. This should eliminate errors if there is a high interrupt load.